Hide progress bar implementation
