+++ libmamba/include/mamba/core/output.hpp
@@ -102,6 +102,8 @@ namespace mamba
         ~ConsoleStream();
     };
 
+    class ProgressBarManager;
+
     class Console
     {
     public:
